make UI for choosing route
