+++ app/build.gradle
@@ -58,6 +58,13 @@ dependencies {
     //permission 을 편리하게 체크할 수 있는 라이브러리
     implementation 'gun0912.ted:tedpermission:2.2.3'
 
+    //navigation component
+    def nav_version = "2.3.1"
+    // Java language implementation
+    implementation "androidx.navigation:navigation-fragment:$nav_version"
+    implementation "androidx.navigation:navigation-ui:$nav_version"
+
+
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
     implementation "androidx.core:core-ktx:+"
