--- app/src/main/java/com/wtm/anshime/ui/main/ChooseRouteFragment.java
@@ -1,23 +1,0 @@
-package com.wtm.anshime.ui.main;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import com.wtm.anshime.R;
-
-public class ChooseRouteFragment extends Fragment {
-
-    public ChooseRouteFragment(){}
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_choose_route, container, false);
-    }
-}
