--- app/src/main/java/com/wtm/anshime/ui/main/WriteInfoFragment.java
@@ -1,30 +1,0 @@
-package com.wtm.anshime.ui.main;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.wtm.anshime.R;
-
-public class WriteInfoFragment extends Fragment {
-    public WriteInfoFragment() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_write_info, container, false);
-    }
-}
