+++ app/src/main/res/drawable/round_corner_rect.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <stroke android:width="1dp" android:color="@color/lightGrey" />
+    <corners android:radius="6dp"/>
+    <solid android:color="@android:color/white"/>
+</shape>
