+++ app/src/main/res/layout/dropdown_menu_popup_item.xml
@@ -1,0 +1,9 @@
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:padding="16dp"
+    android:ellipsize="end"
+    android:maxLines="1"
+    tools:text="도보/자전거"
+    android:textAppearance="?attr/textAppearanceSubtitle1" />
