+++ app/src/main/res/values/colors.xml
@@ -7,5 +7,6 @@
     <color name="kakaoGold">#FFDC00</color>
     <color name="kakaoBlack">#2B291C</color>
 
-    <color name="lightGrey">#E8E8E8</color>
+    <color name="lightGrey">#D1D1D1</color>
+    <color name="greyWhite">#66E6E6E6</color>
     </resources>
