+++ app/src/main/res/values/dimens.xml
@@ -3,4 +3,8 @@
     <dimen name="default_start_margin">20dp</dimen>
     <dimen name="default_end_margin">20dp</dimen>
     <dimen name="card_corner_radius">8dp</dimen>
+    <dimen name="route_input_height">45dp</dimen>
+    <dimen name="route_info_top_margin">200dp</dimen>
+    <dimen name="view_default_margin">12dp</dimen>
+    <dimen name="default_input_height">50dp</dimen>
 </resources>
