+++ app/src/main/res/values/strings.xml
@@ -7,13 +7,24 @@
     <string name="current_location_state_number">%d명</string>
     <string name="arrived_state_title">도착현황 ✨</string>
     <string name="arrived_list_text">**님이 %d분 전에 도착</string>
-    <string name="location_permission_granted_msg">위치권한 허가됨!</string>
-    <string name="location_permission_denied_msg">위치권한 거부됨!</string>
-    <string name="wrong_gps_value_msg">올바른 GPS값이 아니네요ㅠㅠ\n위치 정보를 읽어오지 못했어요!</string>
+    <string name="location_permission_granted_msg">위치권한이 허가되었습니다 :)</string>
+    <string name="location_permission_denied_msg">위치권한이 거부되었습니다 :(</string>
+    <string name="wrong_gps_value_msg">올바른 GPS값이 아닙니다.\n위치 정보를 읽어올 수 없습니다.</string>
     <string name="si_placeholder">\???시</string>
     <string name="main_board_sub_title">여성 안전 귀가 현황</string>
     <string name="main_board_title_placeholder">서울시</string>
     <string name="main_board_total_count">%d명</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="write_route_info_app_bar_title">귀가 정보 작성하기</string>
+    <string name="starting_location">출발지</string>
+    <string name="end_location">도착지</string>
+    <string name="choose_route_next_btn_text">귀가 세부정보 작성하기</string>
+    <string name="write_route_detail_title">귀가 세부정보 작성하기</string>
+    <string name="start_time">출발예상시간</string>
+    <string name="route_info">귀가 정보</string>
+    <string name="end_time">도착예상시간</string>
+    <string name="start_icon_desc_car_number">car</string>
+    <string name="enter_car_number_hint">차량번호 입력</string>
+    <string name="choose_transport_method_hint">이동수단 선택</string>
+    <string name="start_icon_desc_transport_method">transport method</string>
+    <string name="check_info_title">작성정보 확인</string>
 </resources>
